Refactor axios
