+++ src/services/category.service.ts
@@ -1,0 +1,14 @@
+import axios from "axios";
+
+import type { Category } from "@models/category";
+
+import { appConfig } from "@config/app.config";
+
+const getCategories = async (): Promise<Category[]> => {
+  const { data } = await axios.get<Category[]>(
+    `${appConfig.backendUrl}/categories`,
+  );
+  return data;
+};
+
+export { getCategories };
